website/src/backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==2.3.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.0.2
fonttools==4.53.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
pandas==2.0.0
numpy==1.24.0
packaging==24.1
pillow==10.4.0
pulp==2.6.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
pymysql==1.0.2
flask_migrate==4.0.7
scikit-learn==1.2.0
